`sappho-ast`: split out recursive expr types into individual mods.
