+++ ast/src/recursive.rs
@@ -4,27 +4,20 @@
 //! a pure context contains pure expressions, while a list expression in a proc context contains proc
 //! expressions.
 
-use crate::{GenExpr, Pattern};
+mod apply;
+mod letexpr;
+
+use crate::GenExpr;
 use std::fmt;
+
+pub use self::apply::Application;
+pub use self::letexpr::LetExpr;
 
 #[derive(Debug, PartialEq)]
 pub enum RecursiveExpr<Effects> {
     List(Vec<GenExpr<Effects>>),
     Let(LetExpr<Effects>),
     Apply(Application<Effects>),
-}
-
-#[derive(Debug, PartialEq)]
-pub struct LetExpr<Effects> {
-    pub binding: Pattern,
-    pub bindexpr: Box<GenExpr<Effects>>,
-    pub tail: Box<GenExpr<Effects>>,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct Application<Effects> {
-    pub target: Box<GenExpr<Effects>>,
-    pub argument: Box<GenExpr<Effects>>,
 }
 
 impl<FX> fmt::Display for RecursiveExpr<FX>
@@ -52,34 +45,5 @@ where
             Let(x) => x.fmt(f),
             Apply(x) => x.fmt(f),
         }
-    }
-}
-
-impl<FX> fmt::Display for LetExpr<FX>
-where
-    FX: fmt::Display,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "let ")?;
-        self.binding.fmt(f)?;
-        write!(f, " = ")?;
-        self.bindexpr.fmt(f)?;
-        write!(f, "; ")?;
-        self.tail.fmt(f)?;
-        Ok(())
-    }
-}
-
-impl<FX> fmt::Display for Application<FX>
-where
-    FX: fmt::Display,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "(")?;
-        self.target.fmt(f)?;
-        write!(f, " ")?;
-        self.argument.fmt(f)?;
-        write!(f, ")")?;
-        Ok(())
     }
 }
