+++ ast/src/recursive/apply.rs
@@ -1,0 +1,22 @@
+use crate::GenExpr;
+use std::fmt;
+
+#[derive(Debug, PartialEq)]
+pub struct Application<Effects> {
+    pub target: Box<GenExpr<Effects>>,
+    pub argument: Box<GenExpr<Effects>>,
+}
+
+impl<FX> fmt::Display for Application<FX>
+where
+    FX: fmt::Display,
+{
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "(")?;
+        self.target.fmt(f)?;
+        write!(f, " ")?;
+        self.argument.fmt(f)?;
+        write!(f, ")")?;
+        Ok(())
+    }
+}
